recipe-server/requirements/default.txt
bcrypt==3.1.0 \
    --hash=sha256:7d403d74183eb957f34ca331a5d05d4475d7993042419dfd3bfd3f8eb40b98d1 \
    --hash=sha256:d33967627e5f7d777b0e1bccf7aab2587058517d18ddc5b388b2652902511203 \
    --hash=sha256:e6b3c90908a740ce9cc0a0d7d55ae38602e64b226a0658581ea481c822c82fcf \
    --hash=sha256:d43dda44ae1c3d50eddd54029ae1e3e2a2a2734d7e71eb94c66a72f640b1ac4b \
    --hash=sha256:26bad7f8010ccb029990f1103fb3686cdde098408de4855a2a286b07b96aef83 \
    --hash=sha256:e54820d8b9eff357d1003f5b8d4b949a632b76b89610d8a933783fd476033ebe
dj-database-url==0.3.0 \
    --hash=sha256:f2e273ed34acbb560962d5cf12917936d8df02297df09bd3089b8546d4584138 \
    --hash=sha256:ca01768fdecde134301f3170743226f60edff5c3935f12437378ebd911506353
dj-inmemorystorage==1.4.0 \
    --hash=sha256:35a37fac5955c3d9519fcab0cc7d88b43dc9bc95ffd2a6312badfa489ed30fd2
django==1.9.11 \
    --hash=sha256:bec3e58d9458e3121180adf9c33dedae0091ef6e73f79b2f9a0f8c0a34925429 \
    --hash=sha256:dadcfd8b03dfbf472c2d88c12202d9d015af68fb6561099992bc2d91aeab7d9d
django-configurations==1.0 \
    --hash=sha256:98a14951cd0870d0343fe9c243172dc110353b430c3e06a87a90dc7bc82b0b75 \
    --hash=sha256:93f4a09134145f4388c2ec2183ce623b1b57fd04e3e0d2a2cf3b32fadc8f9464
django-countries==3.4.1 \
    --hash=sha256:5bdda9d2c3473b519371428d88517f29befad7e35dfc489e8b0f95cb2aa941dc \
    --hash=sha256:23c6b5455a2e68ed02601cee0d3c80481965d0c3a6bd2f07ca56902b0a4c55a6
django-dirtyfields==0.8.2 \
    --hash=sha256:90a60223a373ec563c0e6b013121a2d965233498384af0cf132b1622d0d8aa19
django-filter==1.0.1 \
    --hash=sha256:03ab2895e086aa579110a8945d7d9c5d28c4b57060cfc697cb5c74a18021fec4 \
    --hash=sha256:7cca9dab22c72df734ba24fb81ca018fdb6d2020c840e758d473dcedd341aa7b
django-mozilla-product-details==0.11.1 \
    --hash=sha256:6a9bf2898deae722876bd19eb874bfdcbe884f99c8096c2ddd6cdea8b95d4fa0
django-npm==0.1.4 \
    --hash=sha256:4ace771ef76bf6a9430278d0f23a23d1a7083ec2b00240c4400d3c5a5cd7cea6
django-reversion==2.0.4 \
    --hash=sha256:a95624e28e72ef0b07231e601abe6fc790cd5e9e54791a1e57ff5eb2b1c0bc32
django-storages-redux==1.3.2 \
    --hash=sha256:df94a582d452c8e7763291d5463b60316bf7f6ffe2c59c4c162514c1bbc39504 \
    --hash=sha256:84cb0e685ac0401f14d320d3d469b3e5968a7314a93936c86672c2315d8302dc
django-sslserver==0.18 \
    --hash=sha256:e7180255546c6e1e1652de032d212aa63b19a26b2c817b5f7babca24711f03ad
django-webpack-loader==0.3.0 \
    --hash=sha256:729c8d1b0d3f1449a2f7964c7e5351d0494bbe0470afdf9570e26593ac75fe13
djangorestframework==3.5.3 \
    --hash=sha256:f446041a944723e14502a0a5880d0bc74a499ac1075781177f2fa6d7fe7b415d \
    --hash=sha256:7c04e2a45e6d30df7042749b9a029882c449eb2fee038f3cc14dfbc93581bfbf
ecdsa==0.13 \
    --hash=sha256:40d002cf360d0e035cf2cb985e1308d41aaa087cbfc135b2dc2d844296ea546c \
    --hash=sha256:64cf1ee26d1cde3c73c6d7d107f835fed7c6a2904aef9eac223d57ad800c43fa
factory-boy==2.6.0 \
    --hash=sha256:116610a170918df342db2c4bafbc8f4b91780c49878f929d89e360f10ad29d2a \
    --hash=sha256:75e4c9786ed28d19ec7fb500f3c7221a6eb87748cbc6f5e2eeaca4c3d68f30cb
flake8==3.3.0 \
    --hash=sha256:83905eadba99f73fbfe966598aaf1682b3eb6755d2263c5b33a4e8367d60b0d1 \
    --hash=sha256:b907a26dcf5580753d8f80f1be0ec1d5c45b719f7bac441120793d1a70b03f12
flake8-junit-report==2.0.1 \
    --hash=sha256:55914a99126d3b30cad2a042d95d466b5b368c7f69ef68ec7ec8321d3f0e2eee
geoip2==2.2.0 \
    --hash=sha256:f362d7300f8a6b968722f5891141faacc009b9acd79fcb9812e43a2c881bff7e \
    --hash=sha256:105ca36e460be58931c162e63b866da10bb29afd22e21b0378bef16086ace853
gevent==1.2.1 \
    --hash=sha256:fa493697197912937cd4a0038a495421dcb066b622294fd95bcd77afa3c8b9ca \
    --hash=sha256:b81ff56d2b3287157f6b0bc8eef3cddb87ee955ae8f7953cf0052b8ef8236ee9 \
    --hash=sha256:c3e414820dca7fa7c14cf002f8d8ea0e41f600bedb2e00146fdd4b3998c6411f \
    --hash=sha256:6b80ee2d4e72521cf600ab5457d4f3269c0c4773305a49571378bfce81098a35 \
    --hash=sha256:3de300d0e32c31311e426e4d5d73b36777ed99c2bac3f8fbad939eeb2c29fa7c \
    --hash=sha256:5e10fb0b97eb749a77b07bc240afc5239e8e34f244033ca3814898213c587ae5 \
    --hash=sha256:0df5f8e95e65c21d4b451874e209c8b016d0f0bd325d0dcce071a95b73e84f43
gunicorn==19.4.5 \
    --hash=sha256:c57f1b005a4b90933303c8deed9bedeb509331aa6a0a990023a5796e52bd8988 \
    --hash=sha256:53b58044764ad79d732af18c580b1a54b724adf4d290ec19c4ca78ab22a1ee0d
hashin==0.4.2 \
    --hash=sha256:4675a60ccb93b4bf2b3625dcd4617b9f7bdb64c6dd463b4a2d39ea6cc03815d3 \
    --hash=sha256:5da3d92f8698ad8c2906b14a421af160014daf55f1d352cac081898ed47aa96a
jsonschema==2.5.1 \
    --hash=sha256:71e7b3bcf9fca408bcb65bb60892f375d3abdd2e4f296eeeb8fe0bbbfcde598e \
    --hash=sha256:36673ac378feed3daa5956276a829699056523d7961027911f064b52255ead41 \
    --hash=sha256:9088494da4c74497a7a27842ae4ca9c3355b5f7754121edc440463eaf020f079
mozilla-cloud-services-logger==1.0.1 \
    --hash=sha256:2a2a524f9d05bc5c8a6f3d13a28cb06386b0720cee00ba58e22b2de68089c6e4
psycopg2==2.6.1 \
    --hash=sha256:6acf9abbbe757ef75dc2ecd9d91ba749547941abaffbe69ff2086a9e37d4904c
pytest-django==2.9.1 \
    --hash=sha256:8be15b637738c8cbd1422a6461465c0aeab7839cf76ad2b5d190b6f1f53facd6 \
    --hash=sha256:743d0056e127ef424850ea76d93d45c92c313da0e56765806a59fc7680c25ab7
pytest==3.0.2 \
    --hash=sha256:4b0872d00159dd8d7a27c4a45a2be77aac8a6e70c3af9a7c76c040c3e3715b9d \
    --hash=sha256:64d8937626dd2a4bc15ef0edd307d26636a72a3f3f9664c424d78e40efb1e339
pytest-mock==0.11.0 \
    --hash=sha256:a897c6283aa116c161f5f08b8f73301685f6b8abfcd0b9b0d000668bc0faaaa3 \
    --hash=sha256:5d109071ac1c6a5a883aece92c507b4562efce86054ab39c83a9f3554688289d
pytest-testrail==0.0.9 \
    --hash=sha256:74a46e3f61eecccb43b55e44a7c99ae22bdf75f85ecf10caf88f03e2a95d0eb3
raven==5.10.2 \
    --hash=sha256:f17269486e9f64ccd5a38a859e8191a8d8f5609e8f0efee6f1881e8a0fb20ed8 \
    --hash=sha256:58d52e16c834f70046f07d53795cafbf2709aef87afa9ca2a9d21be28a6ed92c
requests==2.11.1 \
    --hash=sha256:545c4855cd9d7c12671444326337013766f4eea6068c3f0307fb2dc2696d580e
requests-hawk==1.0.0 \
    --hash=sha256:aef0dff8053dcae2057774516386bed0a3bc03fabea9e18f3aa98f02672ea5d0 \
    --hash=sha256:c2626ab31ebef0c81b97781c44c2275bfcc6d8e8520fc4ced495f0f386f8fe26
Sphinx==1.3.5 \
    --hash=sha256:6b69f27a11085366fedad25bb3c38fe70639712a7e4a234a257ea2ad1f8f66a1 \
    --hash=sha256:b7d133bb4990d010a2ad934c319b52d8a2156cb0491484f5e2a558619bc9ae04
sphinx_rtd_theme==0.1.9 \
    --hash=sha256:3c38d037713bd78043486eea5bf771d71ed697ec25c09e16f49e44887f7fe184 \
    --hash=sha256:46c2ca8b482b0398b9621bb5a8ec7c9c5c7226926e5363e18d744f5f0d39d78f \
    --hash=sha256:273846f8aacac32bf9542365a593b495b68d8035c2e382c9ccedcac387c9a0a1
sphinxcontrib-httpdomain==1.4.0 \
    --hash=sha256:9db07b047b8799378ae532d6b9b17080035da8d889082ad7f124de0469feeea4 \
    --hash=sha256:f4926bfc55e37f09bf6b2759eae20edb92696111a7bf7c39e5e48a1646818866
statsd==3.2.1 \
    --hash=sha256:7aff40c6cdda703193fac6231410c6c0ad5939ce053a168016501e982c755c84 \
    --hash=sha256:3fa92bf0192af926f7a0d9be031fe3fd0fbaa1992d42cf2f07e68f76ac18288e
whitenoise==2.0.6 \
    --hash=sha256:826ffe5d608c9dc8daebef1b0b43d01f7958f17c2fce36e75c80e26160172c4f \
    --hash=sha256:5aea935dfc09ef2beeb76960b4a808b0bbe65e85fb0b0312434b9c365ca02a41
newrelic==2.78.0.57 \
    --hash=sha256:2191b7699e14a07efa5d9221270eb29bdf6cce643aa56cff08546ecb3f729be6
django-csp==3.2 \
    --hash=sha256:7cec78ba7c426deba6d4bea188dcfc1a8b2609ad98dd539c365605c9ec2996b2 \
    --hash=sha256:fb56c0126b4e69f854fd1581a4bef25fa7cac1f10235e95f697e297ceef88132
